+++ packages/backend-app-api/src/services/implementations/httpRouterService.ts
@@ -24,6 +24,16 @@ import Router from 'express-promise-router';
 import { Handler } from 'express';
 import { createServiceBuilder } from '@backstage/backend-common';
 
+/**
+ * @public
+ */
+export type HttpRouterFactoryOptions = {
+  /**
+   * The plugin ID used for the index route. Defaults to 'app'
+   */
+  indexPlugin?: string;
+};
+
 /** @public */
 export const httpRouterFactory = createServiceFactory({
   service: httpRouterServiceRef,
@@ -31,21 +41,28 @@ export const httpRouterFactory = createServiceFactory({
     config: configServiceRef,
     plugin: pluginMetadataServiceRef,
   },
-  async factory({ config }) {
+  async factory({ config }, options?: HttpRouterFactoryOptions) {
+    const defaultPluginId = options?.indexPlugin ?? 'app';
+
+    const apiRouter = Router();
     const rootRouter = Router();
 
     const service = createServiceBuilder(module)
       .loadConfig(config)
+      .addRouter('/api', apiRouter)
       .addRouter('', rootRouter);
 
     await service.start();
 
     return async ({ plugin }) => {
       const pluginId = plugin.getId();
-      const path = pluginId ? `/api/${pluginId}` : '';
       return {
         use(handler: Handler) {
-          rootRouter.use(path, handler);
+          if (pluginId === defaultPluginId) {
+            rootRouter.use(handler);
+          } else {
+            apiRouter.use(`/${pluginId}`, handler);
+          }
         },
       };
     };