+++ packages/backend-next/src/index.ts
@@ -17,9 +17,11 @@
 import { catalogPlugin } from '@backstage/plugin-catalog-backend';
 import { scaffolderCatalogModule } from '@backstage/plugin-scaffolder-backend';
 import { createBackend } from '@backstage/backend-defaults';
+import { appPlugin } from '@backstage/plugin-app-backend';
 
 const backend = createBackend();
 
 backend.add(catalogPlugin());
 backend.add(scaffolderCatalogModule());
+backend.add(appPlugin({ appPackageName: 'example-app' }));
 backend.start();