+++ plugins/catalog-backend-module-openapi/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-catalog-backend-module-openapi
 
+## 0.1.3-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.2-next.0
+  - @backstage/plugin-catalog-backend@1.4.1-next.0
+  - @backstage/plugin-catalog-node@1.1.1-next.0
+  - @backstage/backend-common@0.15.2-next.0
+  - @backstage/config@1.0.3-next.0
+  - @backstage/integration@1.3.2-next.0
+  - @backstage/types@1.0.0
+
 ## 0.1.2
 
 ### Patch Changes